♻️ rename catalogue
