+++ apps/web/components/page/editor/Middle.tsx
@@ -6,7 +6,7 @@ import ChangePositionComponent from "../../base/ChangePositionComponent";
 import ResizeComponent from "../../base/ResizeComponent";
 
 function Middle(props: any) {
-  const { Elements, setIsElement, currentElement, setCurrentElement, pageBackgroundStyle  } =
+  const { Elements, setIsElement, currentElement, setCurrentElement, pageBackgroundStyle } =
     UseElementStore();
 
   return (
@@ -21,7 +21,13 @@ function Middle(props: any) {
       <div
         id="mid-container"
         className="bg-white mt-5"
-        style={{ ...pageBackgroundStyle,width: "375px", height: "667px", position: "relative", overflow: "auto" }}
+        style={{
+          ...pageBackgroundStyle,
+          width: "375px",
+          height: "667px",
+          position: "relative",
+          overflow: "auto",
+        }}
       >
         {Elements.map((item: any) =>
           item.id == currentElement ? (
