+++ apps/web/components/page/index-old/Banner.tsx
