+++ apps/web/components/page/index/Hero.tsx
